got a payment option added to database
