+++ bangazon/ProductManagement.py
@@ -1,4 +1,5 @@
 import sys
+import sqlite3
 
 def get_product(product_name, price):
 	return 3
@@ -26,7 +27,7 @@ def select_payment_option(selected_payment):
 	with sqlite3.connect('bangazon.db') as connection:
 		c = connection.cursor()
 
-		c.execute("insert into order values(?, ?, ?, ?)",
+		c.execute("insert into Order values(?, ?, ?, ?)",
 				(None, payment_type, account_number, active_customer))
 
 		connection.commit()
